Made 'rule' a 'class' using go's methods. Moved totals back to int to make the rule defs easier to write.

+++ automaton_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 type testpair struct {
 	array    []bool
-	expected byte
+	expected int
 }
 
 var tests = []testpair{
